+++ src/bpInfo.ts
@@ -1,1 +1,1 @@
-export const bpInfo = {version: 'v1.16.1'};
+export const bpInfo = {version: 'v1.16.2'};